requirements-v2.txt
# V2 Core Dependencies

# Config & Validation
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP/Async
aiohttp>=3.9.0
aiohttp-retry>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aioresponses>=0.7.0

# Development
black>=23.12.0
mypy>=1.8.0
ruff>=0.1.0
